python-based-API/Payment-template/requirement.txt
fastapi
uvicorn
python-dotenv
stripe
requests
